{routes-spotify-merged} building the new structure with app.js
