+++ api/usernames/username.router.js
@@ -38,6 +38,4 @@ router.get("/", async (req, res) => {
     }
 });
 
-
-
 module.exports = router;
